requirements.txt
openai
pinecone-client


